requirements.txt
python-can
cantools
influxdb